requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==41.0.5
    # via -r requirements.in
django==4.2.4
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-ckeditor
    #   django-debug-toolbar
    #   django-js-asset
    #   djangorestframework
    #   easy-thumbnails
django-appconf==1.0.5
    # via django-compressor
django-ckeditor==6.7.1
    # via -r requirements.in
django-compressor==4.4
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via -r requirements.in
django-js-asset==2.1.0
    # via django-ckeditor
djangorestframework==3.14.0
    # via -r requirements.in
easy-thumbnails==2.8.5
    # via -r requirements.in
factory-boy==3.3.0
    # via -r requirements.in
faker==26.1.0
    # via factory-boy
gunicorn==21.2.0
    # via -r requirements.in
idna==3.4
    # via requests
packaging==23.1
    # via gunicorn
pillow==10.0.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
psycopg2==2.9.7
    # via -r requirements.in
pycparser==2.21
    # via cffi
python-dateutil==2.9.0.post0
    # via faker
pytz==2023.3.post1
    # via djangorestframework
rcssmin==1.1.1
    # via django-compressor
requests==2.31.0
    # via -r requirements.in
rjsmin==1.2.1
    # via django-compressor
sentry-sdk==1.30.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
urllib3==2.0.4
    # via
    #   requests
    #   sentry-sdk
